chore(okms): rename secret manager guides
ref: #MANAGER-19143
