+++ packages/manager/apps/okms/src/modules/secret-manager/components/guides/guide-kv2-api/guideKv2.constants.ts
