+++ packages/manager/apps/okms/src/modules/secret-manager/components/guides/guide-manager/guideManager.constants.ts
