+++ packages/manager/apps/okms/src/modules/secret-manager/components/guides/guide-rest-api/useGuideItemRestApi.ts
