+++ packages/manager/apps/okms/src/modules/secret-manager/pages/okms-dashboard/OkmsDashboard.page.tsx
@@ -21,7 +21,7 @@ import {
 import { queryClient } from '@ovh-ux/manager-react-core-application';
 import { useOkmsById } from '@key-management-service/data/hooks/useOkms';
 import { okmsQueryKeys } from '@key-management-service/data/api/okms';
-import { SecretManagerGuidesButton } from '@secret-manager/components/secret-manager-guides/SecretManagerGuideButton';
+import { SecretManagerGuidesButton } from '@secret-manager/components/guides/SecretManagerGuideButton';
 import { PageSpinner } from '@/common/components/page-spinner/PageSpinner.component';
 import { OkmsDomainDashboardTiles } from '@/common/components/okms-dashboard/OkmsDomainDashboardTiles.component';
 import { OkmsDashboardOutletContext } from './OkmsDashboard.type';
