+++ packages/manager/apps/okms/src/modules/secret-manager/pages/secret/Secret.page.tsx
@@ -21,7 +21,7 @@ import {
   SecretBreadcrumbItem,
 } from '@secret-manager/components/breadcrumb';
 import { SecretManagerChangelogButton } from '@secret-manager/components/secret-manager-changelog-button/SecretManagerChangelogButton.component';
-import { SecretManagerGuidesButton } from '@secret-manager/components/secret-manager-guides/SecretManagerGuideButton';
+import { SecretManagerGuidesButton } from '@secret-manager/components/guides/SecretManagerGuideButton';
 import Loading from '@/common/components/loading/Loading';
 import { SecretPageOutletContext } from './Secret.type';
 import { decodeSecretPath } from '@/modules/secret-manager/utils/secretPath';
